+++ app/ydoc-shared/src/ast/tree.ts
@@ -14,7 +14,7 @@ import type { SpanMap } from './idMap'
 import { newExternalId } from './idMap'
 import type { Module } from './mutableModule'
 import { MutableModule, ROOT_ID } from './mutableModule'
-import { parseExpression, parseStatement } from './parse'
+import { parseBlockStatement, parseExpression, parseModuleStatement } from './parse'
 import type { RawConcreteChild } from './print'
 import {
   ensureSpaced,
@@ -1535,7 +1535,7 @@ export class Import extends BaseStatement {
 
   /** TODO: Add docs */
   static tryParse(source: string, module?: MutableModule): Owned<MutableImport> | undefined {
-    const parsed = parseStatement(source, module)
+    const parsed = parseModuleStatement(source, module)
     if (parsed instanceof MutableImport) return parsed
   }
 
@@ -2036,7 +2036,7 @@ export class ExpressionStatement extends BaseStatement {
     source: string,
     module?: MutableModule,
   ): Owned<MutableExpressionStatement> | undefined {
-    const parsed = parseStatement(source, module)
+    const parsed = parseBlockStatement(source, module)
     if (parsed instanceof MutableExpressionStatement) return parsed
   }
 
@@ -2451,7 +2451,7 @@ export class FunctionDef extends BaseStatement {
 
   /** TODO: Add docs */
   static tryParse(source: string, module?: MutableModule): Owned<MutableFunctionDef> | undefined {
-    const parsed = parseStatement(source, module)
+    const parsed = parseModuleStatement(source, module)
     if (parsed instanceof MutableFunctionDef) return parsed
   }
 
@@ -2764,7 +2764,7 @@ export class Assignment extends BaseStatement {
 
   /** TODO: Add docs */
   static tryParse(source: string, module?: MutableModule): Owned<MutableAssignment> | undefined {
-    const parsed = parseStatement(source, module)
+    const parsed = parseBlockStatement(source, module)
     if (parsed instanceof MutableAssignment) return parsed
   }